added graphics for scalar fields

+++ kernel.c
@@ -1,5 +1,6 @@
 #include "kernel.h"
 #include "accessor.h"
+#include <stdio.h>
 
 #define DEF_KERNEL(KERNEL, TYPE) \
 \
@@ -10,6 +11,6 @@ TYPE *KERNEL##_at(int i, int j, int k) { \
 } \
 
 DEF_KERNEL(kernel_vec, Vector)
-DEF_KERNEL(kernel_scalar_x, float)
-DEF_KERNEL(kernel_scalar_y, float)
-DEF_KERNEL(kernel_scalar_z, float)
+DEF_KERNEL(kernel_scalar_x, double)
+DEF_KERNEL(kernel_scalar_y, double)
+DEF_KERNEL(kernel_scalar_z, double)
